feat: implement `client` for local wallet example
